Split inferences.ts file into submodule for readability
